pms/requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.3
django-monthfield==0.1.3
django-phonenumber-field==5.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
mysqlclient==2.0.1
numpy==1.19.4
phonenumbers==8.12.14
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
